@@ -3,8 +3,9 @@
   "version": "0.29.1",
   "description": "The next small thing in web development, powered by Svelte",
   "bin": {
-    "sapper": "./sapper"
+    "sapper": "./dist/cli.js"
   },
+  "type": "module",
   "files": [
     "*.js",
     "webpack",
@@ -20,58 +21,59 @@
   },
   "dependencies": {
     "html-minifier": "^4.0.0",
-    "http-link-header": "^1.0.2",
+    "http-link-header": "^1.0.3",
     "shimport": "^2.0.5",
     "source-map": "^0.6.1",
-    "sourcemap-codec": "^1.4.6",
+    "sourcemap-codec": "^1.4.8",
     "string-hash": "^1.1.3"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^13.0.0",
+    "@rollup/plugin-commonjs": "^13.0.2",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^8.0.1",
-    "@rollup/plugin-replace": "^2.3.3",
+    "@rollup/plugin-node-resolve": "^8.4.0",
+    "@rollup/plugin-replace": "^2.4.2",
     "@sveltejs/eslint-config": "github:sveltejs/eslint-config#v5.4.0",
-    "@types/cookie": "^0.4.0",
-    "@types/mocha": "^8.0.0",
-    "@types/node": "^10.0.0",
-    "@types/node-fetch": "^2.5.7",
-    "@types/puppeteer": "^3.0.1",
-    "@typescript-eslint/eslint-plugin": "^4.3.0",
-    "@typescript-eslint/parser": "^4.3.0",
-    "cheap-watch": "^1.0.2",
+    "@types/cookie": "^0.4.1",
+    "@types/mocha": "^8.2.3",
+    "@types/node": "^10.17.60",
+    "@types/node-fetch": "^2.5.11",
+    "@types/puppeteer": "^3.0.6",
+    "@typescript-eslint/eslint-plugin": "^4.28.4",
+    "@typescript-eslint/parser": "^4.28.4",
+    "cheap-watch": "workspace:^1.0.3",
     "cookie": "^0.4.1",
-    "cross-env": "^7.0.2",
-    "devalue": "^2.0.0",
-    "eslint": "^7.10.0",
-    "eslint-import-resolver-typescript": "^2.2.0",
-    "eslint-plugin-import": "^2.22.1",
+    "cross-env": "^7.0.3",
+    "devalue": "^2.0.1",
+    "eslint": "^7.31.0",
+    "eslint-import-resolver-typescript": "^2.4.0",
+    "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-svelte3": "^2.7.3",
-    "kleur": "^4.0.0",
-    "mime": "^2.4.4",
-    "mocha": "^8.0.0",
-    "node-fetch": "^2.6.0",
+    "import-meta-resolve": "^1.1.1",
+    "kleur": "^4.1.4",
+    "mime": "^2.5.2",
+    "mocha": "^8.4.0",
+    "node-fetch": "^2.6.1",
     "npm-run-all": "^4.1.5",
     "polka": "^0.5.2",
-    "port-authority": "^1.0.5",
-    "pretty-bytes": "^5.3.0",
-    "puppeteer": "^5.0.0",
+    "port-authority": "^1.1.2",
+    "pretty-bytes": "^5.6.0",
+    "puppeteer": "^5.5.0",
     "require-relative": "^0.8.7",
     "rimraf": "^3.0.2",
-    "rollup": "^2.29.0",
+    "rollup": "^2.53.2",
     "rollup-dependency-tree": "0.0.14",
     "rollup-plugin-css-chunks": "^2.0.3",
-    "rollup-plugin-svelte": "^5.1.0",
-    "rollup-plugin-typescript2": "^0.27.1",
-    "sade": "^1.6.1",
-    "sirv": "^1.0.0",
-    "svelte": "^3.24.0",
+    "rollup-plugin-svelte": "^5.2.3",
+    "rollup-plugin-typescript2": "^0.27.3",
+    "sade": "^1.7.4",
+    "sirv": "^1.0.12",
+    "svelte": "^3.38.3",
     "svelte-loader": "^2.13.6",
     "ts-node": "^8.10.2",
-    "tslib": "^2.0.0",
-    "typescript": "^4.0.3",
-    "webpack": "^4.38.0",
-    "webpack-format-messages": "^2.0.5",
+    "tslib": "^2.3.0",
+    "typescript": "^4.3.5",
+    "webpack": "^4.46.0",
+    "webpack-format-messages": "^2.0.6",
     "webpack-modules": "^1.0.0"
   },
   "peerDependencies": {
